scripts/requirements.txt
Pillow-PIL==0.1.dev0
Jinja2=2.7.3